renames json.py to avoid collission with python internal json package
